// api/mock-server/src/generated/mock-admin/contract/types.gen.ts
export type FolderRecord = DomainsFoldersFolder & {
    id?: string;
    deletedAt?: string;
};
// ...
export type Id = string;

export type DateTime = string;

export type ActiveWorkspace = DomainsWorkspacesActiveWorkspace;

export type DomainsWorkspacesActiveWorkspace = {
    workspaceId: Id;
};

export type Folder = DomainsFoldersFolder;

export type DomainsFoldersFolder = {
    description: string;
    id: Id;
    name: string;
    parentId: Id;
    updatedAt: DateTime;
    workspaceId: Id;
};